+++ modules/dxp/apps/commerce/commerce-shipping-engine-fixed-service/src/main/java/com/liferay/commerce/shipping/engine/fixed/service/impl/CommerceShippingFixedOptionRelLocalServiceImpl.java
@@ -21,6 +21,7 @@ import com.liferay.portal.kernel.model.User;
 import com.liferay.portal.kernel.service.ServiceContext;
 import com.liferay.portal.kernel.util.OrderByComparator;
 
+import java.math.BigDecimal;
 import java.util.List;
 
 /**
@@ -34,8 +35,9 @@ public class CommerceShippingFixedOptionRelLocalServiceImpl
 			long commerceShippingMethodId, long commerceShippingFixedOptionId,
 			long commerceWarehouseId, long commerceCountryId,
 			long commerceRegionId, String zip, double weightFrom,
-			double weightTo, double fixedPrice, double rateUnitWeightPrice,
-			double ratePercentage, ServiceContext serviceContext)
+			double weightTo, BigDecimal fixedPrice,
+			BigDecimal rateUnitWeightPrice, double ratePercentage,
+			ServiceContext serviceContext)
 		throws PortalException {
 
 		User user = userLocalService.getUser(serviceContext.getUserId());
@@ -156,8 +158,8 @@ public class CommerceShippingFixedOptionRelLocalServiceImpl
 	public CommerceShippingFixedOptionRel updateCommerceShippingFixedOptionRel(
 			long commerceShippingFixedOptionRelId, long commerceWarehouseId,
 			long commerceCountryId, long commerceRegionId, String zip,
-			double weightFrom, double weightTo, double fixedPrice,
-			double rateUnitWeightPrice, double ratePercentage)
+			double weightFrom, double weightTo, BigDecimal fixedPrice,
+			BigDecimal rateUnitWeightPrice, double ratePercentage)
 		throws PortalException {
 
 		CommerceShippingFixedOptionRel commerceShippingFixedOptionRel =
